+++ app/helpers/navigation_helper.rb
@@ -7,8 +7,13 @@ module NavigationHelper
     link_to title, talks_path
   end
 
-  def goals_link
-    link_to t('goals.plural'), goals_path
+  # Skip until 1.1 version
+  # def goals_link
+  #   link_to t('goals.plural'), goals_path
+  # end
+
+  def donate_link
+    link_to t('donations.donate'), donate_path
   end
 
   def chat_link
