+++ app/views/donates/show.slim
@@ -1,0 +1,3 @@
+h2 = t 'donations.donate'
+
+= render 'shared/donations/form', path: donate_path
