+++ app/views/layouts/app/_header.slim
@@ -6,7 +6,8 @@ nav
   ul
     li = events_link
     li = talks_link
-    li = goals_link
+    li = donate_link
+    / li = goals_link # Skip until 1.1 version
     li = chat_link
     li = members_link
     li = speakers_link
